Fix JSNapi::NumberRef Description   1. Finding key 1 in Map ([[1, "CC"]]) returns undefined.   2. Fix the problem in JSNapi::NumberRef, the complete constructor is not defined in this interface. Issue:   #I541HH:Fix JSNapi::NumberRef

+++ ecmascript/tooling/base/pt_params.cpp
@@ -758,7 +758,7 @@ Local<ObjectRef> StartSamplingParams::ToObject(const EcmaVM *ecmaVm)
     Local<ObjectRef> params = NewObject(ecmaVm);
 
     params->Set(ecmaVm, Local<JSValueRef>(StringRef::NewFromUtf8(ecmaVm, "samplingInterval")),
-            NumberRef::New(ecmaVm, samplingInterval_.value()));
+            NumberRef::New(ecmaVm, static_cast<int64_t>(samplingInterval_.value())));
 
     return params;
 }
